Motor control for rotation and index
Now it's possible to control the rotation and the position of the targets. It's necessary to calibrate the position of each target and check how to get the index position to get that as a reference at the initialization of the control software.  I think we need to use the function HM or HI.